+++ app/models/sheet.rb
@@ -416,11 +416,37 @@ class Sheet < ActiveRecord::Base
     (self.response_count * 100.0 / self.total_response_count).to_i rescue 0
   end
 
+
+  def non_hidden_variable_ids
+    @non_hidden_variable_ids ||= begin
+      variable_ids = []
+      self.design.options.each do |option|
+        if variable = Variable.current.find_by_id(option[:variable_id]) and self.show_variable?(option[:branching_logic])
+          variable_ids << variable.id
+        end
+      end
+      variable_ids
+    end
+  end
+
+  def non_hidden_responses
+    non_blank_sheet_variable_ids = self.sheet_variables.collect{|sv| sv.empty_or_not}.compact
+    sheet_variables = SheetVariable.where( variable_id: self.non_hidden_variable_ids ).where( id: non_blank_sheet_variable_ids ).count
+  end
+
+  def non_hidden_total_responses
+    self.non_hidden_variable_ids.count
+  end
+
   def check_response_count_change
     if self.total_design_variables != self.total_response_count
-      self.update_column :response_count, self.non_blank_design_variable_responses
-      self.update_column :total_response_count, self.total_design_variables
+      self.update_response_count!
     end
+  end
+
+  def update_response_count!
+    self.update_column :response_count, self.non_hidden_responses
+    self.update_column :total_response_count, self.non_hidden_total_responses
   end
 
   def coverage